+++ exm-web/components/ui/image.tsx
@@ -16,7 +16,7 @@ const Image: FC<
     src,
     fill = true,
     alt = "",
-    onError = () => setSrcI(props.fallBack || "/product.png"),
+    onError = () => setSrcI(props.fallBack || "/images/error.png"),
     width,
     height,
     fallBack,