+++ exm-web/modules/feed/component/form/EventForm.tsx
@@ -1,14 +1,19 @@
 "use client"
 
+import "react-datetime-picker/dist/DateTimePicker.css"
+import "react-calendar/dist/Calendar.css"
+import "react-clock/dist/Clock.css"
 import { useEffect, useState } from "react"
 import { zodResolver } from "@hookform/resolvers/zod"
-import { XCircle } from "lucide-react"
+import dayjs from "dayjs"
+import relativeTime from "dayjs/plugin/relativeTime"
+import { Calendar } from "lucide-react"
+import DateTimePicker from "react-datetime-picker"
 import { useForm } from "react-hook-form"
 import Select from "react-select"
 import * as z from "zod"
 
 import { Button } from "@/components/ui/button"
-import { DatePicker } from "@/components/ui/date-picker"
 import {
   DialogContent,
   DialogHeader,
@@ -23,18 +28,24 @@ import {
   FormMessage,
 } from "@/components/ui/form"
 import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
 import LoadingPost from "@/components/ui/loadingPost"
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group"
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from "@/components/ui/popover"
 import SuccessPost from "@/components/ui/successPost"
 import { Textarea } from "@/components/ui/textarea"
-import { AttachmentWithPreview } from "@/components/AttachmentWithPreview"
 import SelectUsers from "@/components/select/SelectUsers"
 
 import useFeedMutation from "../../hooks/useFeedMutation"
 import { useTeamMembers } from "../../hooks/useTeamMembers"
 import { IFeed } from "../../types"
+import FormAttachments from "./FormAttachments"
+import FormImages from "./FormImages"
 import Uploader from "./uploader/Uploader"
+
+dayjs.extend(relativeTime)
 
 const FormSchema = z
   .object({
@@ -64,7 +75,7 @@ const FormSchema = z
     departmentIds: z.array(z.string()).optional(),
     branchIds: z.array(z.string()).optional(),
     unitId: z.string().optional(),
-    recipientIds: z.array(z.string()).optional(),
+    recipientIds: z.array(z.object({})).optional(),
   })
   .refine(
     ({ startDate, endDate }) => {
@@ -108,6 +119,7 @@ const EventForm = ({
   const [imageUploading, setImageUploading] = useState(false)
   const [attachmentUploading, setAttachmentUploading] = useState(false)
   const [recipientIds, setRecipientIds] = useState(feed?.recipientIds || [])
+  const [value, setValue] = useState(new Date())
 
   const callBack = (result: string) => {
     if (result === "success") {
@@ -181,22 +193,6 @@ const EventForm = ({
     )
   }
 
-  const deleteAttachment = (index: number) => {
-    const updated = [...attachments]
-
-    updated.splice(index, 1)
-
-    setAttachments(updated)
-  }
-
-  const deleteImage = (index: number) => {
-    const updated = [...images]
-
-    updated.splice(index, 1)
-
-    setImage(updated)
-  }
-
   const onChangeMultiValue = (type: string, datas: any) => {
     const onchangeFunc = type === "department" ? setDepartmentIds : setBranchIds
 
@@ -206,7 +202,7 @@ const EventForm = ({
   }
 
   return (
-    <DialogContent className="max-h-[110vh] overflow-auto">
+    <DialogContent className="max-h-[80vh] max-w-2xl overflow-auto">
       <DialogHeader>
         <DialogTitle>Create event</DialogTitle>
       </DialogHeader>
@@ -214,21 +210,16 @@ const EventForm = ({
       {mutationLoading ? <LoadingPost /> : null}
       {success ? <SuccessPost /> : null}
 
-      <RadioGroup
-        className="flex font-semibold"
-        value={visibility}
-        onValueChange={(value) => setVisibility(value)}
-      >
-        <div>
-          <RadioGroupItem value="public" id="1" className="mr-1" />
-          <Label htmlFor="1">Public</Label>
+      <div className="flex justify-end items-center">
+        <div
+          className="cursor-pointer px-2 py-1 hover:bg-[#F0F0F0] rounded-md"
+          onClick={() =>
+            setVisibility(visibility === "public" ? "private" : "public")
+          }
+        >
+          {visibility === "private" ? "Make public" : "Make private"}
         </div>
-
-        <div>
-          <RadioGroupItem value="private" id="2" className="mr-1" />
-          <Label htmlFor="2">Private</Label>
-        </div>
-      </RadioGroup>
+      </div>
 
       <Form {...form}>
         <form className="space-y-3" onSubmit={form.handleSubmit(onSubmit)}>
@@ -252,51 +243,85 @@ const EventForm = ({
             />
           )}
 
-          <FormField
-            control={form.control}
-            name="startDate"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel className="block">Start Date</FormLabel>
-                <FormControl>
-                  <DatePicker
-                    date={field.value}
-                    setDate={field.onChange}
-                    fromDate={new Date()}
-                    className="w-full"
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
+          <div className="flex gap-3">
+            <FormField
+              control={form.control}
+              name="startDate"
+              render={({ field }) => (
+                <FormItem className="flex-1">
+                  <FormControl>
+                    <Popover>
+                      <PopoverTrigger className="w-full">
+                        <div className="cursor-pointer flex rounded-md px-3 py-2 border items-center w-full border-light">
+                          <Calendar size={15} className="mr-2" />
+                          <span>
+                            {field.value
+                              ? dayjs(field.value).format("MM/DD/YYYY HH:mm")
+                              : "Start date"}
+                          </span>
+                        </div>
+                      </PopoverTrigger>
+                      <PopoverContent className="w-fit p-2">
+                        <DateTimePicker
+                          format="yyyy/MM/dd h:mm:ss a"
+                          value={field.value}
+                          onChange={field.onChange}
+                          minDate={new Date()}
+                          clockClassName="hidden"
+                        />
+                      </PopoverContent>
+                    </Popover>
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
 
-          <FormField
-            control={form.control}
-            name="endDate"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel className="block">End Date</FormLabel>
-                <FormControl>
-                  <DatePicker
-                    date={field.value}
-                    setDate={field.onChange}
-                    className="w-full"
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
+            <FormField
+              control={form.control}
+              name="endDate"
+              render={({ field }) => (
+                <FormItem className="flex-1">
+                  <FormControl>
+                    <Popover>
+                      <PopoverTrigger className="w-full">
+                        <div className="cursor-pointer flex rounded-md px-3 py-2 border items-center w-full border-light">
+                          <Calendar size={15} className="mr-2" />
+                          <span>
+                            {field.value
+                              ? dayjs(field.value).format("MM/DD/YYYY HH:mm")
+                              : "End date"}
+                          </span>
+                        </div>
+                      </PopoverTrigger>
+                      <PopoverContent className="w-fit p-2">
+                        <DateTimePicker
+                          format="yyyy/MM/dd h:mm:ss a"
+                          value={field.value}
+                          onChange={field.onChange}
+                          minDate={new Date()}
+                          clockClassName="hidden"
+                        />
+                      </PopoverContent>
+                    </Popover>
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+          </div>
 
           <FormField
             control={form.control}
             name="title"
             render={({ field }) => (
               <FormItem>
-                <FormLabel>Title</FormLabel>
                 <FormControl>
-                  <Input placeholder="title" {...field} />
+                  <Input
+                    placeholder="Event name"
+                    {...field}
+                    className="border-light"
+                  />
                 </FormControl>
                 <FormMessage />
               </FormItem>
@@ -308,12 +333,12 @@ const EventForm = ({
             name="description"
             render={({ field }) => (
               <FormItem>
-                <FormLabel>Description</FormLabel>
                 <FormControl>
                   <Textarea
-                    placeholder="description"
+                    placeholder="Event description"
                     {...field}
                     defaultValue={feed?.description || ""}
+                    className="border-light"
                   />
                 </FormControl>
                 <FormMessage />
@@ -326,11 +351,10 @@ const EventForm = ({
             name="where"
             render={({ field }) => (
               <FormItem>
-                <FormLabel>Where</FormLabel>
                 <FormControl>
                   <Textarea
-                    className="rounded-md px-3 py-2"
-                    placeholder="where"
+                    className="rounded-md px-3 py-2 border-light"
+                    placeholder="Add location"
                     {...field}
                     defaultValue={feed?.eventData?.where || ""}
                   />
@@ -344,7 +368,6 @@ const EventForm = ({
             name="departmentIds"
             render={() => (
               <FormItem>
-                <FormLabel>Departments</FormLabel>
                 <FormControl>
                   {loading && !reload && !seDepartmentSearchvalue ? (
                     <Input disabled={true} placeholder="Loading..." />
@@ -358,7 +381,7 @@ const EventForm = ({
                         (departmentOption) =>
                           departmentIds.includes(departmentOption?.value)
                       )}
-                      placeholder="Select departments"
+                      placeholder="All departments"
                       isSearchable={true}
                       onInputChange={seDepartmentSearchvalue}
                       onChange={(data) =>
@@ -377,7 +400,6 @@ const EventForm = ({
             name="branchIds"
             render={() => (
               <FormItem>
-                <FormLabel>Branches</FormLabel>
                 <FormControl>
                   {loading && !reload && branchSearchValue ? (
                     <Input disabled={true} placeholder="Loading..." />
@@ -390,7 +412,7 @@ const EventForm = ({
                       defaultValue={branchOptions?.filter((branchOption) =>
                         branchIds?.includes(branchOption?.value)
                       )}
-                      placeholder="Select branches"
+                      placeholder="All branches"
                       isSearchable={true}
                       onInputChange={setBranchSearchvalue}
                       onChange={(data) => onChangeMultiValue("branch", data)}
@@ -407,7 +429,6 @@ const EventForm = ({
             name="unitId"
             render={({}) => (
               <FormItem>
-                <FormLabel>Unit</FormLabel>
                 <FormControl>
                   {loading && !reload && unitSearchValue ? (
                     <Input disabled={true} placeholder="Loading..." />
@@ -417,7 +438,7 @@ const EventForm = ({
                       onMenuOpen={() => setReload(true)}
                       isClearable={true}
                       options={unitOptions}
-                      placeholder="Select units"
+                      placeholder="All units"
                       value={unitOptions?.filter(
                         (unitOption) => unitOption.value === unitId
                       )}
@@ -434,37 +455,33 @@ const EventForm = ({
             )}
           />
 
-          <Uploader
-            defaultFileList={images || []}
-            onChange={setImage}
-            type={"image"}
-            setUploading={setImageUploading}
+          <FormAttachments
+            attachments={attachments || []}
+            setAttachments={setImage}
           />
-          {images && images.length > 0 && (
-            <AttachmentWithPreview
-              images={images}
-              className="mt-2"
-              deleteImage={deleteImage}
-            />
-          )}
+          <FormImages images={images || []} setImage={setImage} />
 
-          <Uploader
-            defaultFileList={attachments || []}
-            onChange={setAttachments}
-            setUploading={setAttachmentUploading}
-          />
+          <div className="flex items-center border rounded-lg px-2 border-[#cccccc] justify-between">
+            <p className="text-[#444]">Add attachments</p>
+            <div className="flex">
+              <Uploader
+                defaultFileList={images || []}
+                onChange={setImage}
+                type={"image"}
+                icon={true}
+                iconSize={20}
+                setUploading={setImageUploading}
+              />
 
-          {(attachments || []).map((attachment, index) => {
-            return (
-              <div
-                key={index}
-                className="flex items-center border-y text-sm font-semibold text-[#444] p-2.5"
-              >
-                {attachment.name}{" "}
-                <XCircle size={18} onClick={() => deleteAttachment(index)} />
-              </div>
-            )
-          })}
+              <Uploader
+                defaultFileList={attachments || []}
+                onChange={setAttachments}
+                icon={true}
+                iconSize={20}
+                setUploading={setAttachmentUploading}
+              />
+            </div>
+          </div>
 
           <Button
             type="submit"