+++ packages/plugin-goals-api/src/utils.ts
@@ -1,4 +1,8 @@
 import { sendCommonMessage } from './messageBroker';
+import { sendCoreMessage } from './messageBroker';
+import { generateModels, IModels } from './connectionResolver';
+import { IUserDocument } from '@erxes/api-utils/src/types';
+import { IGoalDocument } from './models/definitions/goals';
 
 export const countDocuments = async (
   subdomain: string,