+++ packages/plugin-loyalties-ui/src/App.tsx
@@ -3,6 +3,9 @@ import GeneralRoutes from './generalRoutes';
 import { PluginLayout } from '@erxes/ui/src/styles/main';
 import { AppProvider } from 'coreui/appContext';
 
+import '@erxes/ui/src/styles/global-styles';
+import 'erxes-icon/css/erxes.min.css';
+
 const App = () => {
   return (
     <AppProvider>