+++ pos/types/slots.type.ts
@@ -1,8 +1,9 @@
 export interface ISlot {
   _id: string
+  posToken: string
   code: string
   name: string
-  isPreDates: string | null
+  isPreDates: { dueDate: string; _id: string }[]
   option: {
     width: number
     height: number